Unit selection might work, but not tested in car yet.
