+++ Apps/PcmLogger/PcmLogger.csproj
@@ -85,6 +85,9 @@
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="MainForm.Can.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="PathDisplayAdapter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
